+++ src/features/admin/hooks/index.ts
@@ -6,6 +6,7 @@ export * from "./useAdminCategories";
 export * from "./useAdminCoupons";
 export * from "./useAdminFaqs";
 export * from "./useAdminNewsletter";
+export * from "./useAdminOrders";
 export * from "./useAdminPayouts";
 export * from "./useAdminProducts";
 export * from "./useAdminReviews";